added conformation before deleteing car

+++ src/Catalog/Car/Car.js
@@ -15,25 +15,28 @@ export default class Car extends Component {
 
     handleClickDelete = e => {
         e.preventDefault()
-        const carId = this.props.car.id
-        fetch(`${config.API_ENDPOINT}/cars/${carId}`, {
-            method: 'DELETE',
-            headers: {
-                'content-type': 'application/json'
-            },
-        })
-            .then(res => {
-                if (!res.ok)
-                    return res.json().then(e => Promise.reject(e))
-                return res
+        let confirmDelete = window.confirm('Are you sure you want to delete car?')
+        if (confirmDelete) {
+            const carId = this.props.car.id
+            fetch(`${config.API_ENDPOINT}/cars/${carId}`, {
+                method: 'DELETE',
+                headers: {
+                    'content-type': 'application/json'
+                },
             })
-            .then(() => {
-                this.context.removeCar(carId)
-                this.props.history.push('/catalog')
-            })
-            .catch(error => {
-                console.error({ error })
-            })
+                .then(res => {
+                    if (!res.ok)
+                        return res.json().then(e => Promise.reject(e))
+                    return res
+                })
+                .then(() => {
+                    this.context.removeCar(carId)
+                    this.props.history.push('/catalog')
+                })
+                .catch(error => {
+                    console.error({ error })
+                })
+        }
     }
     render() {
         /* const carImages = (this.props.carImages.map((img, i) => {
@@ -45,8 +48,8 @@ export default class Car extends Component {
                 <span className='make car-display'>Make: {(this.props.car ? this.props.car.make : null)}</span>
                 <span className='year car-display'>Year: {(this.props.car ? this.props.car.year : null)}</span>
                 <span className='desc car-display'>Description: {(this.props.car ? this.props.car.description : null)}</span>
-                {(this.props.car.manufacturer !== ('[object Object]') ? <span className='manufacturer car-display'>Manufacturer: {this.props.car.manufacturer}</span>: <></>) }
-                {(this.props.car.scale !== ('[object Object]') ? <span className='scale car-display'>Scale: {this.props.car.scale }</span> : <></>)}
+                {(this.props.car.manufacturer !== ('[object Object]') ? <span className='manufacturer car-display'>Manufacturer: {this.props.car.manufacturer}</span> : <></>)}
+                {(this.props.car.scale !== ('[object Object]') ? <span className='scale car-display'>Scale: {this.props.car.scale}</span> : <></>)}
                 <button
                     className='car__delete'
                     type='button'
